test(tau-coding-agent): split commands_and_packages test modules (#1379)
