+++ crates/tau-coding-agent/src/tests/auth_provider/commands_and_packages.rs
@@ -5698,1902 +5698,6 @@ fn unit_validate_event_webhook_ingest_cli_accepts_signed_github_configuration() 
     validate_event_webhook_ingest_cli(&cli).expect("signed github webhook config should pass");
 }
 
-#[test]
-fn functional_execute_channel_store_admin_inspect_succeeds() {
-    let temp = tempdir().expect("tempdir");
-    let store = crate::channel_store::ChannelStore::open(temp.path(), "github", "issue-1")
-        .expect("open channel store");
-    store
-        .append_log_entry(&crate::channel_store::ChannelLogEntry {
-            timestamp_unix_ms: 1,
-            direction: "inbound".to_string(),
-            event_key: Some("e1".to_string()),
-            source: "github".to_string(),
-            payload: serde_json::json!({"body":"hello"}),
-        })
-        .expect("append log");
-    store
-        .write_text_artifact(
-            "run-active",
-            "github-reply",
-            "private",
-            Some(30),
-            "md",
-            "artifact body",
-        )
-        .expect("write artifact");
-    let mut artifact_index =
-        std::fs::read_to_string(store.artifact_index_path()).expect("read artifact index");
-    artifact_index.push_str("invalid-artifact-line\n");
-    std::fs::write(store.artifact_index_path(), artifact_index).expect("seed invalid artifact");
+mod channel_store_admin_reports;
 
-    let mut cli = test_cli();
-    cli.channel_store_root = temp.path().to_path_buf();
-    cli.channel_store_inspect = Some("github/issue-1".to_string());
-
-    execute_channel_store_admin_command(&cli).expect("inspect should succeed");
-    let report = store.inspect().expect("inspect report");
-    assert_eq!(report.artifact_records, 1);
-    assert_eq!(report.invalid_artifact_lines, 1);
-    assert_eq!(report.active_artifacts, 1);
-    assert_eq!(report.expired_artifacts, 0);
-}
-
-#[test]
-fn regression_execute_channel_store_admin_repair_removes_invalid_lines() {
-    let temp = tempdir().expect("tempdir");
-    let store = crate::channel_store::ChannelStore::open(temp.path(), "slack", "C123")
-        .expect("open channel store");
-    std::fs::write(store.log_path(), "{\"ok\":true}\ninvalid-json-line\n")
-        .expect("seed invalid log");
-    let expired = store
-        .write_text_artifact(
-            "run-expired",
-            "slack-reply",
-            "private",
-            Some(0),
-            "md",
-            "expired artifact",
-        )
-        .expect("write expired artifact");
-    let mut artifact_index =
-        std::fs::read_to_string(store.artifact_index_path()).expect("read artifact index");
-    artifact_index.push_str("invalid-artifact-line\n");
-    std::fs::write(store.artifact_index_path(), artifact_index).expect("seed invalid artifact");
-
-    let mut cli = test_cli();
-    cli.channel_store_root = temp.path().to_path_buf();
-    cli.channel_store_repair = Some("slack/C123".to_string());
-    execute_channel_store_admin_command(&cli).expect("repair should succeed");
-
-    let report = store.inspect().expect("inspect after repair");
-    assert_eq!(report.invalid_log_lines, 0);
-    assert_eq!(report.log_records, 1);
-    assert_eq!(report.invalid_artifact_lines, 0);
-    assert_eq!(report.expired_artifacts, 0);
-    assert_eq!(report.active_artifacts, 0);
-    assert!(!store.channel_dir().join(expired.relative_path).exists());
-}
-
-#[test]
-fn functional_execute_channel_store_admin_github_status_inspect_succeeds() {
-    let temp = tempdir().expect("tempdir");
-    let github_state_dir = temp.path().join("github");
-    let repo_state_dir = github_state_dir.join("owner__repo");
-    std::fs::create_dir_all(&repo_state_dir).expect("create github repo dir");
-    std::fs::write(
-        repo_state_dir.join("state.json"),
-        r#"{
-  "schema_version": 1,
-  "last_issue_scan_at": "2026-01-01T00:00:00Z",
-  "processed_event_keys": ["issue-comment-created:1"],
-  "issue_sessions": {
-    "7": {
-      "session_id": "issue-7",
-      "last_run_id": "run-7",
-      "last_event_key": "issue-comment-created:1",
-      "last_event_kind": "issue_comment_created",
-      "last_actor_login": "alice",
-      "last_reason_code": "command_processed",
-      "total_processed_events": 1
-    }
-  },
-  "health": {
-    "updated_unix_ms": 700,
-    "cycle_duration_ms": 25,
-    "queue_depth": 0,
-    "active_runs": 0,
-    "failure_streak": 0,
-    "last_cycle_discovered": 1,
-    "last_cycle_processed": 1,
-    "last_cycle_completed": 1,
-    "last_cycle_failed": 0,
-    "last_cycle_duplicates": 0
-  }
-}
-"#,
-    )
-    .expect("write github state");
-    std::fs::write(
-        repo_state_dir.join("inbound-events.jsonl"),
-        r#"{"kind":"issue_comment_created","event_key":"issue-comment-created:1"}
-"#,
-    )
-    .expect("write inbound");
-    std::fs::write(
-        repo_state_dir.join("outbound-events.jsonl"),
-        r#"{"event_key":"issue-comment-created:1","command":"chat-status","status":"reported","reason_code":"command_processed"}
-"#,
-    )
-    .expect("write outbound");
-
-    let mut cli = test_cli();
-    cli.github_status_inspect = Some("owner/repo".to_string());
-    cli.github_status_json = true;
-    cli.github_state_dir = github_state_dir;
-
-    execute_channel_store_admin_command(&cli).expect("github status inspect should succeed");
-}
-
-#[test]
-fn regression_execute_channel_store_admin_github_status_inspect_requires_state_file() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    cli.github_status_inspect = Some("owner/repo".to_string());
-    cli.github_state_dir = temp.path().join("github");
-    std::fs::create_dir_all(cli.github_state_dir.join("owner__repo")).expect("create repo dir");
-
-    let error = execute_channel_store_admin_command(&cli)
-        .expect_err("github status inspect should fail without state file");
-    assert!(error.to_string().contains("failed to read"));
-    assert!(error.to_string().contains("state.json"));
-}
-
-#[test]
-fn functional_execute_channel_store_admin_dashboard_status_inspect_succeeds() {
-    let temp = tempdir().expect("tempdir");
-    let dashboard_state_dir = temp.path().join("dashboard");
-    std::fs::create_dir_all(&dashboard_state_dir).expect("create dashboard state dir");
-    std::fs::write(
-        dashboard_state_dir.join("state.json"),
-        r#"{
-  "schema_version": 1,
-  "processed_case_keys": ["snapshot:s1"],
-  "widget_views": [{"widget_id":"health-summary"}],
-  "control_audit": [{"case_id":"c1"}],
-  "health": {
-    "updated_unix_ms": 700,
-    "cycle_duration_ms": 20,
-    "queue_depth": 0,
-    "active_runs": 0,
-    "failure_streak": 0,
-    "last_cycle_discovered": 1,
-    "last_cycle_processed": 1,
-    "last_cycle_completed": 1,
-    "last_cycle_failed": 0,
-    "last_cycle_duplicates": 0
-  }
-}
-"#,
-    )
-    .expect("write dashboard state");
-    std::fs::write(
-        dashboard_state_dir.join("runtime-events.jsonl"),
-        r#"{"reason_codes":["widget_views_updated"],"health_reason":"no recent transport failures observed"}
-"#,
-    )
-    .expect("write dashboard events");
-
-    let mut cli = test_cli();
-    cli.dashboard_status_inspect = true;
-    cli.dashboard_status_json = true;
-    cli.dashboard_state_dir = dashboard_state_dir;
-    execute_channel_store_admin_command(&cli).expect("dashboard status inspect should succeed");
-}
-
-#[test]
-fn regression_execute_channel_store_admin_dashboard_status_inspect_requires_state_file() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    cli.dashboard_status_inspect = true;
-    cli.dashboard_state_dir = temp.path().join("dashboard");
-    std::fs::create_dir_all(&cli.dashboard_state_dir).expect("create dashboard dir");
-
-    let error = execute_channel_store_admin_command(&cli)
-        .expect_err("dashboard status inspect should fail without state file");
-    assert!(error.to_string().contains("failed to read"));
-    assert!(error.to_string().contains("state.json"));
-}
-
-#[test]
-fn functional_execute_channel_store_admin_multi_channel_status_inspect_succeeds() {
-    let temp = tempdir().expect("tempdir");
-    let multi_channel_state_dir = temp.path().join("multi-channel");
-    std::fs::create_dir_all(&multi_channel_state_dir).expect("create multi-channel state dir");
-    std::fs::write(
-        multi_channel_state_dir.join("state.json"),
-        r#"{
-  "schema_version": 1,
-  "processed_event_keys": ["telegram:tg-1", "discord:dc-1", "whatsapp:wa-1"],
-  "health": {
-    "updated_unix_ms": 701,
-    "cycle_duration_ms": 16,
-    "queue_depth": 0,
-    "active_runs": 0,
-    "failure_streak": 0,
-    "last_cycle_discovered": 3,
-    "last_cycle_processed": 3,
-    "last_cycle_completed": 3,
-    "last_cycle_failed": 0,
-    "last_cycle_duplicates": 0
-  }
-}
-"#,
-    )
-    .expect("write multi-channel state");
-    std::fs::write(
-        multi_channel_state_dir.join("runtime-events.jsonl"),
-        r#"{"reason_codes":["healthy_cycle","events_applied"],"health_reason":"no recent transport failures observed"}
-"#,
-    )
-    .expect("write multi-channel events");
-
-    let mut cli = test_cli();
-    cli.multi_channel_status_inspect = true;
-    cli.multi_channel_status_json = true;
-    cli.multi_channel_state_dir = multi_channel_state_dir;
-    execute_channel_store_admin_command(&cli).expect("multi-channel status inspect should succeed");
-}
-
-#[test]
-fn regression_execute_channel_store_admin_multi_channel_status_inspect_requires_state_file() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    cli.multi_channel_status_inspect = true;
-    cli.multi_channel_state_dir = temp.path().join("multi-channel");
-    std::fs::create_dir_all(&cli.multi_channel_state_dir).expect("create multi-channel dir");
-
-    let error = execute_channel_store_admin_command(&cli)
-        .expect_err("multi-channel status inspect should fail without state file");
-    assert!(error.to_string().contains("failed to read"));
-    assert!(error.to_string().contains("state.json"));
-}
-
-#[test]
-fn functional_build_multi_channel_route_inspect_report_resolves_binding_and_role() {
-    let temp = tempdir().expect("tempdir");
-    let route_table_path = temp.path().join("route-table.json");
-    let state_dir = temp.path().join("multi-channel");
-    let security_dir = state_dir.join("security");
-    std::fs::create_dir_all(&security_dir).expect("create security dir");
-    std::fs::write(
-        route_table_path.as_path(),
-        r#"{
-  "schema_version": 1,
-  "roles": {
-    "triage": {},
-    "default": {}
-  },
-  "planner": { "role": "default" },
-  "delegated": { "role": "default" },
-  "delegated_categories": {
-    "incident": { "role": "triage" }
-  },
-  "review": { "role": "default" }
-}"#,
-    )
-    .expect("write route table");
-    std::fs::write(
-        security_dir.join("multi-channel-route-bindings.json"),
-        r#"{
-  "schema_version": 1,
-  "bindings": [
-    {
-      "binding_id": "discord-ops",
-      "transport": "discord",
-      "account_id": "discord-main",
-      "conversation_id": "ops-room",
-      "actor_id": "*",
-      "phase": "delegated_step",
-      "category_hint": "incident",
-      "session_key_template": "session-{role}"
-    }
-  ]
-}"#,
-    )
-    .expect("write route bindings");
-    let event_path = temp.path().join("event.json");
-    std::fs::write(
-        &event_path,
-        r#"{
-  "schema_version": 1,
-  "transport": "discord",
-  "event_kind": "message",
-  "event_id": "dc-route-1",
-  "conversation_id": "ops-room",
-  "actor_id": "discord-user-1",
-  "timestamp_ms": 1760200000000,
-  "text": "please triage this incident",
-  "metadata": {
-    "account_id": "discord-main"
-  }
-}"#,
-    )
-    .expect("write event file");
-
-    let mut cli = test_cli();
-    cli.multi_channel_route_inspect_file = Some(event_path.clone());
-    cli.multi_channel_state_dir = state_dir;
-    cli.orchestrator_route_table = Some(route_table_path);
-    let report = build_multi_channel_route_inspect_report(
-        &tau_multi_channel::MultiChannelRouteInspectConfig {
-            inspect_file: event_path.clone(),
-            state_dir: cli.multi_channel_state_dir.clone(),
-            orchestrator_route_table_path: cli.orchestrator_route_table.clone(),
-        },
-    )
-    .expect("build report");
-    assert_eq!(report.binding_id, "discord-ops");
-    assert_eq!(report.selected_role, "triage");
-    assert_eq!(report.phase, "delegated-step");
-    assert_eq!(report.session_key, "session-triage");
-}
-
-#[test]
-fn integration_execute_multi_channel_route_inspect_command_accepts_live_envelope_input() {
-    let temp = tempdir().expect("tempdir");
-    let state_dir = temp.path().join("multi-channel");
-    std::fs::create_dir_all(state_dir.join("security")).expect("create security dir");
-    let envelope_path = temp.path().join("telegram-envelope.json");
-    std::fs::write(
-        &envelope_path,
-        r#"{
-  "schema_version": 1,
-  "transport": "telegram",
-  "provider": "telegram-bot-api",
-  "payload": {
-    "update_id": 70001,
-    "message": {
-      "message_id": 501,
-      "date": 1760200100,
-      "text": "/status",
-      "chat": { "id": "chat-100" },
-      "from": { "id": "user-100", "username": "ops-user" }
-    }
-  }
-}"#,
-    )
-    .expect("write envelope file");
-
-    let mut cli = test_cli();
-    cli.multi_channel_route_inspect_file = Some(envelope_path.clone());
-    cli.multi_channel_route_inspect_json = true;
-    cli.multi_channel_state_dir = state_dir;
-    let report = build_multi_channel_route_inspect_report(
-        &tau_multi_channel::MultiChannelRouteInspectConfig {
-            inspect_file: envelope_path.clone(),
-            state_dir: cli.multi_channel_state_dir.clone(),
-            orchestrator_route_table_path: cli.orchestrator_route_table.clone(),
-        },
-    )
-    .expect("build report");
-    serde_json::to_string_pretty(&report).expect("render report");
-}
-
-#[test]
-fn regression_build_multi_channel_route_inspect_report_rejects_empty_file() {
-    let temp = tempdir().expect("tempdir");
-    let event_path = temp.path().join("empty.json");
-    std::fs::write(&event_path, "  \n").expect("write empty event file");
-
-    let mut cli = test_cli();
-    cli.multi_channel_route_inspect_file = Some(event_path.clone());
-    let error = build_multi_channel_route_inspect_report(
-        &tau_multi_channel::MultiChannelRouteInspectConfig {
-            inspect_file: event_path.clone(),
-            state_dir: cli.multi_channel_state_dir.clone(),
-            orchestrator_route_table_path: cli.orchestrator_route_table.clone(),
-        },
-    )
-    .expect_err("empty route inspect file should fail");
-    assert!(error.to_string().contains("is empty"));
-    assert!(error
-        .to_string()
-        .contains(event_path.to_string_lossy().as_ref()));
-}
-
-#[test]
-fn unit_build_multi_channel_incident_timeline_report_aggregates_outcomes_and_reason_codes() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    set_workspace_tau_paths(&mut cli, temp.path());
-    cli.multi_channel_incident_timeline = true;
-
-    let channel_dir = cli
-        .multi_channel_state_dir
-        .join("channel-store/channels/discord/ops-room");
-    std::fs::create_dir_all(&channel_dir).expect("create channel dir");
-    std::fs::write(
-        channel_dir.join("log.jsonl"),
-        r#"{"timestamp_unix_ms":1760200000000,"direction":"inbound","event_key":"evt-allow","source":"discord","payload":{"transport":"discord","conversation_id":"ops-room","route_session_key":"ops-room","route":{"binding_id":"discord-ops","binding_matched":true},"channel_policy":{"reason_code":"allow_channel_policy_allow_from_any"}}}
-{"timestamp_unix_ms":1760200000010,"direction":"outbound","event_key":"evt-allow","source":"tau-multi-channel-runner","payload":{"event_key":"evt-allow","response":"ok","delivery":{"mode":"dry_run","receipts":[{"status":"dry_run"}]}}}
-{"timestamp_unix_ms":1760200000020,"direction":"inbound","event_key":"evt-denied","source":"discord","payload":{"transport":"discord","conversation_id":"ops-room","route_session_key":"ops-room","route":{"binding_id":"discord-ops","binding_matched":true},"channel_policy":{"reason_code":"deny_channel_policy_mention_required"}}}
-{"timestamp_unix_ms":1760200000030,"direction":"outbound","event_key":"evt-denied","source":"tau-multi-channel-runner","payload":{"status":"denied","reason_code":"deny_channel_policy_mention_required"}}
-{"timestamp_unix_ms":1760200000040,"direction":"inbound","event_key":"evt-retried","source":"discord","payload":{"transport":"discord","conversation_id":"ops-room","route_session_key":"ops-room","route":{"binding_id":"discord-ops","binding_matched":true},"channel_policy":{"reason_code":"allow_channel_policy_allow_from_any"}}}
-{"timestamp_unix_ms":1760200000050,"direction":"outbound","event_key":"evt-retried","source":"tau-multi-channel-runner","payload":{"status":"delivery_failed","reason_code":"delivery_provider_unavailable","retryable":true}}
-{"timestamp_unix_ms":1760200000060,"direction":"outbound","event_key":"evt-retried","source":"tau-multi-channel-runner","payload":{"event_key":"evt-retried","response":"after retry","delivery":{"mode":"provider","receipts":[{"status":"sent"}]}}}
-{"timestamp_unix_ms":1760200000070,"direction":"inbound","event_key":"evt-failed","source":"discord","payload":{"transport":"discord","conversation_id":"ops-room","route_session_key":"ops-room","route":{"binding_id":"discord-ops","binding_matched":true},"channel_policy":{"reason_code":"allow_channel_policy_allow_from_any"}}}
-{"timestamp_unix_ms":1760200000080,"direction":"outbound","event_key":"evt-failed","source":"tau-multi-channel-runner","payload":{"status":"delivery_failed","reason_code":"delivery_request_rejected","retryable":false}}
-"#,
-    )
-    .expect("write channel log");
-
-    let report = build_multi_channel_incident_timeline_report(
-        &tau_multi_channel::MultiChannelIncidentTimelineQuery {
-            state_dir: cli.multi_channel_state_dir.clone(),
-            window_start_unix_ms: cli.multi_channel_incident_start_unix_ms,
-            window_end_unix_ms: cli.multi_channel_incident_end_unix_ms,
-            event_limit: cli.multi_channel_incident_event_limit.unwrap_or(200),
-            replay_export_path: cli.multi_channel_incident_replay_export.clone(),
-        },
-    )
-    .expect("build incident timeline report");
-    assert_eq!(report.timeline.len(), 4);
-    assert_eq!(report.outcomes.allowed, 1);
-    assert_eq!(report.outcomes.denied, 1);
-    assert_eq!(report.outcomes.retried, 1);
-    assert_eq!(report.outcomes.failed, 1);
-    assert_eq!(
-        report
-            .policy_reason_code_counts
-            .get("allow_channel_policy_allow_from_any")
-            .copied(),
-        Some(3)
-    );
-    assert_eq!(
-        report
-            .policy_reason_code_counts
-            .get("deny_channel_policy_mention_required")
-            .copied(),
-        Some(1)
-    );
-    assert_eq!(
-        report
-            .delivery_reason_code_counts
-            .get("delivery_provider_unavailable")
-            .copied(),
-        Some(1)
-    );
-    assert_eq!(
-        report
-            .delivery_reason_code_counts
-            .get("delivery_request_rejected")
-            .copied(),
-        Some(1)
-    );
-}
-
-#[test]
-fn functional_build_multi_channel_incident_timeline_report_writes_replay_export() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    set_workspace_tau_paths(&mut cli, temp.path());
-    cli.multi_channel_incident_timeline = true;
-    cli.multi_channel_incident_event_limit = Some(1);
-    let replay_export_path = temp.path().join("artifacts/incident-replay.json");
-    cli.multi_channel_incident_replay_export = Some(replay_export_path.clone());
-
-    let channel_dir = cli
-        .multi_channel_state_dir
-        .join("channel-store/channels/telegram/ops-chat");
-    std::fs::create_dir_all(&channel_dir).expect("create channel dir");
-    std::fs::write(
-        channel_dir.join("log.jsonl"),
-        r#"{"timestamp_unix_ms":1760200100000,"direction":"inbound","event_key":"evt-1","source":"telegram","payload":{"transport":"telegram","conversation_id":"ops-chat","route_session_key":"ops-chat","route":{"binding_id":"telegram-ops","binding_matched":true},"channel_policy":{"reason_code":"allow_channel_policy_allow_from_any"}}}
-{"timestamp_unix_ms":1760200100010,"direction":"outbound","event_key":"evt-1","source":"tau-multi-channel-runner","payload":{"event_key":"evt-1","response":"ok","delivery":{"mode":"dry_run","receipts":[{"status":"dry_run"}]}}}
-{"timestamp_unix_ms":1760200100020,"direction":"inbound","event_key":"evt-2","source":"telegram","payload":{"transport":"telegram","conversation_id":"ops-chat","route_session_key":"ops-chat","route":{"binding_id":"telegram-ops","binding_matched":true},"channel_policy":{"reason_code":"allow_channel_policy_allow_from_any"}}}
-{"timestamp_unix_ms":1760200100030,"direction":"outbound","event_key":"evt-2","source":"tau-multi-channel-runner","payload":{"status":"denied","reason_code":"deny_channel_policy_mention_required"}}
-"#,
-    )
-    .expect("write channel log");
-
-    let report = build_multi_channel_incident_timeline_report(
-        &tau_multi_channel::MultiChannelIncidentTimelineQuery {
-            state_dir: cli.multi_channel_state_dir.clone(),
-            window_start_unix_ms: cli.multi_channel_incident_start_unix_ms,
-            window_end_unix_ms: cli.multi_channel_incident_end_unix_ms,
-            event_limit: cli.multi_channel_incident_event_limit.unwrap_or(200),
-            replay_export_path: cli.multi_channel_incident_replay_export.clone(),
-        },
-    )
-    .expect("build incident timeline report");
-    assert_eq!(report.timeline.len(), 1);
-    assert_eq!(report.truncated_event_count, 1);
-    let replay_export = report.replay_export.expect("replay export summary");
-    assert_eq!(replay_export.path, replay_export_path.display().to_string());
-    assert!(!replay_export.checksum_sha256.is_empty());
-
-    let replay_raw = std::fs::read_to_string(&replay_export_path).expect("read replay artifact");
-    let replay_json: serde_json::Value =
-        serde_json::from_str(&replay_raw).expect("parse replay artifact");
-    assert_eq!(replay_json["schema_version"].as_u64(), Some(1));
-    assert_eq!(
-        replay_json["events"].as_array().map(|items| items.len()),
-        Some(1)
-    );
-}
-
-#[test]
-fn integration_build_multi_channel_incident_timeline_report_reads_sample_channel_store_corpus() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    set_workspace_tau_paths(&mut cli, temp.path());
-    cli.multi_channel_incident_timeline = true;
-
-    let telegram_dir = cli
-        .multi_channel_state_dir
-        .join("channel-store/channels/telegram/tg-ops");
-    let discord_dir = cli
-        .multi_channel_state_dir
-        .join("channel-store/channels/discord/dc-ops");
-    let whatsapp_dir = cli
-        .multi_channel_state_dir
-        .join("channel-store/channels/whatsapp/wa-ops");
-    std::fs::create_dir_all(&telegram_dir).expect("create telegram dir");
-    std::fs::create_dir_all(&discord_dir).expect("create discord dir");
-    std::fs::create_dir_all(&whatsapp_dir).expect("create whatsapp dir");
-    std::fs::write(
-        telegram_dir.join("log.jsonl"),
-        r#"{"timestamp_unix_ms":1760200400000,"direction":"inbound","event_key":"evt-tg","source":"telegram","payload":{"transport":"telegram","conversation_id":"tg-ops","route_session_key":"tg-ops","route":{"binding_id":"tg-ops","binding_matched":true},"channel_policy":{"reason_code":"allow_channel_policy_allow_from_any"}}}
-{"timestamp_unix_ms":1760200400010,"direction":"outbound","event_key":"evt-tg","source":"tau-multi-channel-runner","payload":{"event_key":"evt-tg","response":"ok","delivery":{"mode":"dry_run","receipts":[{"status":"dry_run"}]}}}
-"#,
-    )
-    .expect("write telegram log");
-    std::fs::write(
-        discord_dir.join("log.jsonl"),
-        r#"{"timestamp_unix_ms":1760200400100,"direction":"inbound","event_key":"evt-dc","source":"discord","payload":{"transport":"discord","conversation_id":"dc-ops","route_session_key":"dc-ops","route":{"binding_id":"dc-ops","binding_matched":true},"channel_policy":{"reason_code":"allow_channel_policy_allow_from_any"}}}
-{"timestamp_unix_ms":1760200400110,"direction":"outbound","event_key":"evt-dc","source":"tau-multi-channel-runner","payload":{"status":"delivery_failed","reason_code":"delivery_provider_unavailable","retryable":true}}
-{"timestamp_unix_ms":1760200400120,"direction":"outbound","event_key":"evt-dc","source":"tau-multi-channel-runner","payload":{"event_key":"evt-dc","response":"ok","delivery":{"mode":"provider","receipts":[{"status":"sent"}]}}}
-"#,
-    )
-    .expect("write discord log");
-    std::fs::write(
-        whatsapp_dir.join("log.jsonl"),
-        r#"{"timestamp_unix_ms":1760200400200,"direction":"inbound","event_key":"evt-wa","source":"whatsapp","payload":{"transport":"whatsapp","conversation_id":"wa-ops","route_session_key":"wa-ops","route":{"binding_id":"wa-ops","binding_matched":false},"channel_policy":{"reason_code":"deny_channel_policy_mention_required"}}}
-{"timestamp_unix_ms":1760200400210,"direction":"outbound","event_key":"evt-wa","source":"tau-multi-channel-runner","payload":{"status":"denied","reason_code":"deny_channel_policy_mention_required"}}
-"#,
-    )
-    .expect("write whatsapp log");
-
-    let report = build_multi_channel_incident_timeline_report(
-        &tau_multi_channel::MultiChannelIncidentTimelineQuery {
-            state_dir: cli.multi_channel_state_dir.clone(),
-            window_start_unix_ms: cli.multi_channel_incident_start_unix_ms,
-            window_end_unix_ms: cli.multi_channel_incident_end_unix_ms,
-            event_limit: cli.multi_channel_incident_event_limit.unwrap_or(200),
-            replay_export_path: cli.multi_channel_incident_replay_export.clone(),
-        },
-    )
-    .expect("build incident timeline report");
-    assert_eq!(report.timeline.len(), 3);
-    assert_eq!(report.scanned_channel_count, 3);
-    assert_eq!(report.outcomes.allowed, 1);
-    assert_eq!(report.outcomes.retried, 1);
-    assert_eq!(report.outcomes.denied, 1);
-    assert_eq!(
-        report
-            .route_reason_code_counts
-            .get("route_binding_matched")
-            .copied(),
-        Some(2)
-    );
-    assert_eq!(
-        report
-            .route_reason_code_counts
-            .get("route_binding_default")
-            .copied(),
-        Some(1)
-    );
-}
-
-#[test]
-fn regression_build_multi_channel_incident_timeline_report_tolerates_malformed_log_lines() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    set_workspace_tau_paths(&mut cli, temp.path());
-    cli.multi_channel_incident_timeline = true;
-
-    let channel_dir = cli
-        .multi_channel_state_dir
-        .join("channel-store/channels/whatsapp/ops-room");
-    std::fs::create_dir_all(&channel_dir).expect("create channel dir");
-    std::fs::write(
-        channel_dir.join("log.jsonl"),
-        r#"{"timestamp_unix_ms":1760200200000,"direction":"inbound","event_key":"evt-1","source":"whatsapp","payload":{"transport":"whatsapp","conversation_id":"ops-room","route_session_key":"ops-room","route":{"binding_id":"wa-ops","binding_matched":false},"channel_policy":{"reason_code":"allow_channel_policy_allow_from_any"}}}
-{bad-json-line
-{"timestamp_unix_ms":1760200200010,"direction":"outbound","event_key":"evt-1","source":"tau-multi-channel-runner","payload":{"event_key":"evt-1","response":"ok","delivery":{"mode":"dry_run","receipts":[{"status":"dry_run"}]}}}
-"#,
-    )
-    .expect("write channel log");
-
-    let report = build_multi_channel_incident_timeline_report(
-        &tau_multi_channel::MultiChannelIncidentTimelineQuery {
-            state_dir: cli.multi_channel_state_dir.clone(),
-            window_start_unix_ms: cli.multi_channel_incident_start_unix_ms,
-            window_end_unix_ms: cli.multi_channel_incident_end_unix_ms,
-            event_limit: cli.multi_channel_incident_event_limit.unwrap_or(200),
-            replay_export_path: cli.multi_channel_incident_replay_export.clone(),
-        },
-    )
-    .expect("build incident timeline report");
-    assert_eq!(report.timeline.len(), 1);
-    assert_eq!(report.invalid_line_count, 1);
-    assert!(
-        !report.diagnostics.is_empty(),
-        "malformed line should surface diagnostic message"
-    );
-}
-
-#[test]
-fn functional_execute_channel_store_admin_multi_agent_status_inspect_succeeds() {
-    let temp = tempdir().expect("tempdir");
-    let multi_agent_state_dir = temp.path().join("multi-agent");
-    std::fs::create_dir_all(&multi_agent_state_dir).expect("create multi-agent state dir");
-    std::fs::write(
-        multi_agent_state_dir.join("state.json"),
-        r#"{
-  "schema_version": 1,
-  "processed_case_keys": ["planner:planner-success"],
-  "routed_cases": [
-    {
-      "case_key": "planner:planner-success",
-      "case_id": "planner-success",
-      "phase": "planner",
-      "selected_role": "planner",
-      "attempted_roles": ["planner"],
-      "category": "planning",
-      "updated_unix_ms": 1
-    }
-  ],
-  "health": {
-    "updated_unix_ms": 702,
-    "cycle_duration_ms": 19,
-    "queue_depth": 0,
-    "active_runs": 0,
-    "failure_streak": 0,
-    "last_cycle_discovered": 1,
-    "last_cycle_processed": 1,
-    "last_cycle_completed": 1,
-    "last_cycle_failed": 0,
-    "last_cycle_duplicates": 0
-  }
-}
-"#,
-    )
-    .expect("write multi-agent state");
-    std::fs::write(
-        multi_agent_state_dir.join("runtime-events.jsonl"),
-        r#"{"reason_codes":["routed_cases_updated"],"health_reason":"no recent transport failures observed"}
-"#,
-    )
-    .expect("write multi-agent events");
-
-    let mut cli = test_cli();
-    cli.multi_agent_status_inspect = true;
-    cli.multi_agent_status_json = true;
-    cli.multi_agent_state_dir = multi_agent_state_dir;
-    execute_channel_store_admin_command(&cli).expect("multi-agent status inspect should succeed");
-}
-
-#[test]
-fn regression_execute_channel_store_admin_multi_agent_status_inspect_requires_state_file() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    cli.multi_agent_status_inspect = true;
-    cli.multi_agent_state_dir = temp.path().join("multi-agent");
-    std::fs::create_dir_all(&cli.multi_agent_state_dir).expect("create multi-agent dir");
-
-    let error = execute_channel_store_admin_command(&cli)
-        .expect_err("multi-agent status inspect should fail without state file");
-    assert!(error.to_string().contains("failed to read"));
-    assert!(error.to_string().contains("state.json"));
-}
-
-#[test]
-fn functional_execute_channel_store_admin_gateway_status_inspect_succeeds() {
-    let temp = tempdir().expect("tempdir");
-    let gateway_state_dir = temp.path().join("gateway");
-    std::fs::create_dir_all(&gateway_state_dir).expect("create gateway state dir");
-    std::fs::write(
-        gateway_state_dir.join("state.json"),
-        r#"{
-  "schema_version": 1,
-  "processed_case_keys": ["POST:/v1/tasks:gateway-success"],
-  "requests": [
-    {
-      "case_key": "POST:/v1/tasks:gateway-success",
-      "case_id": "gateway-success",
-      "method": "POST",
-      "endpoint": "/v1/tasks",
-      "actor_id": "ops-bot",
-      "status_code": 201,
-      "outcome": "success",
-      "error_code": "",
-      "response_body": {"status":"accepted"},
-      "updated_unix_ms": 1
-    }
-  ],
-  "health": {
-    "updated_unix_ms": 705,
-    "cycle_duration_ms": 15,
-    "queue_depth": 0,
-    "active_runs": 0,
-    "failure_streak": 0,
-    "last_cycle_discovered": 1,
-    "last_cycle_processed": 1,
-    "last_cycle_completed": 1,
-    "last_cycle_failed": 0,
-    "last_cycle_duplicates": 0
-  }
-}
-"#,
-    )
-    .expect("write gateway state");
-    std::fs::write(
-        gateway_state_dir.join("runtime-events.jsonl"),
-        r#"{"reason_codes":["healthy_cycle"],"health_reason":"no recent transport failures observed"}
-"#,
-    )
-    .expect("write gateway events");
-
-    let mut cli = test_cli();
-    cli.gateway_status_inspect = true;
-    cli.gateway_status_json = true;
-    cli.gateway_state_dir = gateway_state_dir;
-    execute_channel_store_admin_command(&cli).expect("gateway status inspect should succeed");
-}
-
-#[test]
-fn regression_execute_channel_store_admin_gateway_status_inspect_requires_state_file() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    cli.gateway_status_inspect = true;
-    cli.gateway_state_dir = temp.path().join("gateway");
-    std::fs::create_dir_all(&cli.gateway_state_dir).expect("create gateway dir");
-
-    let error = execute_channel_store_admin_command(&cli)
-        .expect_err("gateway status inspect should fail without state file");
-    assert!(error.to_string().contains("failed to read"));
-    assert!(error.to_string().contains("state.json"));
-}
-
-#[test]
-fn functional_execute_channel_store_admin_custom_command_status_inspect_succeeds() {
-    let temp = tempdir().expect("tempdir");
-    let custom_command_state_dir = temp.path().join("custom-command");
-    std::fs::create_dir_all(&custom_command_state_dir).expect("create custom-command state dir");
-    std::fs::write(
-        custom_command_state_dir.join("state.json"),
-        r#"{
-  "schema_version": 1,
-  "processed_case_keys": ["CREATE:deploy_release:create-1"],
-  "commands": [
-    {
-      "case_key": "CREATE:deploy_release:create-1",
-      "case_id": "create-1",
-      "command_name": "deploy_release",
-      "template": "deploy {{env}}",
-      "operation": "CREATE",
-      "last_status_code": 201,
-      "last_outcome": "success",
-      "run_count": 1,
-      "updated_unix_ms": 1
-    }
-  ],
-  "health": {
-    "updated_unix_ms": 710,
-    "cycle_duration_ms": 14,
-    "queue_depth": 0,
-    "active_runs": 0,
-    "failure_streak": 0,
-    "last_cycle_discovered": 1,
-    "last_cycle_processed": 1,
-    "last_cycle_completed": 1,
-    "last_cycle_failed": 0,
-    "last_cycle_duplicates": 0
-  }
-}
-"#,
-    )
-    .expect("write custom-command state");
-    std::fs::write(
-        custom_command_state_dir.join("runtime-events.jsonl"),
-        r#"{"reason_codes":["command_registry_mutated"],"health_reason":"no recent transport failures observed"}
-"#,
-    )
-    .expect("write custom-command events");
-
-    let mut cli = test_cli();
-    cli.custom_command_status_inspect = true;
-    cli.custom_command_status_json = true;
-    cli.custom_command_state_dir = custom_command_state_dir;
-    execute_channel_store_admin_command(&cli)
-        .expect("custom-command status inspect should succeed");
-}
-
-#[test]
-fn regression_execute_channel_store_admin_custom_command_status_inspect_requires_state_file() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    cli.custom_command_status_inspect = true;
-    cli.custom_command_state_dir = temp.path().join("custom-command");
-    std::fs::create_dir_all(&cli.custom_command_state_dir).expect("create custom-command dir");
-
-    let error = execute_channel_store_admin_command(&cli)
-        .expect_err("custom-command status inspect should fail without state file");
-    assert!(error.to_string().contains("failed to read"));
-    assert!(error.to_string().contains("state.json"));
-}
-
-#[test]
-fn functional_execute_channel_store_admin_voice_status_inspect_succeeds() {
-    let temp = tempdir().expect("tempdir");
-    let voice_state_dir = temp.path().join("voice");
-    std::fs::create_dir_all(&voice_state_dir).expect("create voice state dir");
-    std::fs::write(
-        voice_state_dir.join("state.json"),
-        r#"{
-  "schema_version": 1,
-  "processed_case_keys": ["turn:tau:ops-1:voice-success-turn"],
-  "interactions": [
-    {
-      "case_key": "turn:tau:ops-1:voice-success-turn",
-      "case_id": "voice-success-turn",
-      "mode": "turn",
-      "wake_word": "tau",
-      "locale": "en-US",
-      "speaker_id": "ops-1",
-      "utterance": "open dashboard",
-      "last_status_code": 202,
-      "last_outcome": "success",
-      "run_count": 1,
-      "updated_unix_ms": 1
-    }
-  ],
-  "health": {
-    "updated_unix_ms": 720,
-    "cycle_duration_ms": 12,
-    "queue_depth": 0,
-    "active_runs": 0,
-    "failure_streak": 0,
-    "last_cycle_discovered": 1,
-    "last_cycle_processed": 1,
-    "last_cycle_completed": 1,
-    "last_cycle_failed": 0,
-    "last_cycle_duplicates": 0
-  }
-}
-"#,
-    )
-    .expect("write voice state");
-    std::fs::write(
-        voice_state_dir.join("runtime-events.jsonl"),
-        r#"{"reason_codes":["turns_handled"],"health_reason":"no recent transport failures observed"}
-"#,
-    )
-    .expect("write voice events");
-
-    let mut cli = test_cli();
-    cli.voice_status_inspect = true;
-    cli.voice_status_json = true;
-    cli.voice_state_dir = voice_state_dir;
-    execute_channel_store_admin_command(&cli).expect("voice status inspect should succeed");
-}
-
-#[test]
-fn regression_execute_channel_store_admin_voice_status_inspect_requires_state_file() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    cli.voice_status_inspect = true;
-    cli.voice_state_dir = temp.path().join("voice");
-    std::fs::create_dir_all(&cli.voice_state_dir).expect("create voice dir");
-
-    let error = execute_channel_store_admin_command(&cli)
-        .expect_err("voice status inspect should fail without state file");
-    assert!(error.to_string().contains("failed to read"));
-    assert!(error.to_string().contains("state.json"));
-}
-
-#[test]
-fn functional_execute_extension_validate_command_succeeds_for_valid_manifest() {
-    let temp = tempdir().expect("tempdir");
-    let manifest_path = temp.path().join("extension.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "id": "issue-assistant",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "bin/assistant",
-  "hooks": ["run-start", "run-end"],
-  "permissions": ["read-files", "network"],
-  "timeout_ms": 60000
-}"#,
-    )
-    .expect("write extension manifest");
-
-    let mut cli = test_cli();
-    cli.extension_validate = Some(manifest_path);
-    execute_extension_validate_command(&cli).expect("extension validate should succeed");
-}
-
-#[test]
-fn regression_execute_extension_validate_command_rejects_invalid_manifest() {
-    let temp = tempdir().expect("tempdir");
-    let manifest_path = temp.path().join("extension.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "id": "issue-assistant",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "../escape.sh"
-}"#,
-    )
-    .expect("write extension manifest");
-
-    let mut cli = test_cli();
-    cli.extension_validate = Some(manifest_path);
-    let error =
-        execute_extension_validate_command(&cli).expect_err("unsafe entrypoint should fail");
-    assert!(error
-        .to_string()
-        .contains("must not contain parent traversals"));
-}
-
-#[test]
-fn functional_execute_extension_show_command_succeeds_for_valid_manifest() {
-    let temp = tempdir().expect("tempdir");
-    let manifest_path = temp.path().join("extension.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "id": "issue-assistant",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "bin/assistant",
-  "hooks": ["run-end", "run-start"],
-  "permissions": ["network", "read-files"]
-}"#,
-    )
-    .expect("write extension manifest");
-
-    let mut cli = test_cli();
-    cli.extension_show = Some(manifest_path);
-    execute_extension_show_command(&cli).expect("extension show should succeed");
-}
-
-#[test]
-fn regression_execute_extension_show_command_rejects_invalid_manifest() {
-    let temp = tempdir().expect("tempdir");
-    let manifest_path = temp.path().join("extension.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 9,
-  "id": "issue-assistant",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "bin/assistant"
-}"#,
-    )
-    .expect("write extension manifest");
-
-    let mut cli = test_cli();
-    cli.extension_show = Some(manifest_path);
-    let error = execute_extension_show_command(&cli).expect_err("invalid schema should fail");
-    assert!(error
-        .to_string()
-        .contains("unsupported extension manifest schema"));
-}
-
-#[test]
-fn functional_execute_extension_list_command_reports_mixed_inventory() {
-    let temp = tempdir().expect("tempdir");
-    let root = temp.path().join("extensions");
-    let valid_dir = root.join("issue-assistant");
-    std::fs::create_dir_all(&valid_dir).expect("create valid extension dir");
-    std::fs::write(
-        valid_dir.join("extension.json"),
-        r#"{
-  "schema_version": 1,
-  "id": "issue-assistant",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "bin/assistant"
-}"#,
-    )
-    .expect("write valid extension manifest");
-    let invalid_dir = root.join("broken");
-    std::fs::create_dir_all(&invalid_dir).expect("create invalid extension dir");
-    std::fs::write(
-        invalid_dir.join("extension.json"),
-        r#"{
-  "schema_version": 9,
-  "id": "broken",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "bin/assistant"
-}"#,
-    )
-    .expect("write invalid extension manifest");
-
-    let mut cli = test_cli();
-    cli.extension_list = true;
-    cli.extension_list_root = root;
-    execute_extension_list_command(&cli).expect("extension list should succeed");
-}
-
-#[test]
-fn regression_execute_extension_list_command_rejects_non_directory_root() {
-    let temp = tempdir().expect("tempdir");
-    let root_file = temp.path().join("extensions.json");
-    std::fs::write(&root_file, "{}").expect("write root file");
-
-    let mut cli = test_cli();
-    cli.extension_list = true;
-    cli.extension_list_root = root_file;
-    let error =
-        execute_extension_list_command(&cli).expect_err("non-directory extension root should fail");
-    assert!(error.to_string().contains("is not a directory"));
-}
-
-#[test]
-fn functional_execute_extension_exec_command_runs_process_hook() {
-    let temp = tempdir().expect("tempdir");
-    let bin_dir = temp.path().join("bin");
-    std::fs::create_dir_all(&bin_dir).expect("create bin directory");
-    let script_path = bin_dir.join("hook.sh");
-    std::fs::write(
-        &script_path,
-        "#!/usr/bin/env bash\nread -r _input\nprintf '{\"ok\":true,\"result\":\"hook-processed\"}'\n",
-    )
-    .expect("write hook script");
-    make_script_executable(&script_path);
-    let manifest_path = temp.path().join("extension.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "id": "issue-assistant",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "bin/hook.sh",
-  "hooks": ["run-start"],
-  "permissions": ["run-commands"],
-  "timeout_ms": 5000
-}"#,
-    )
-    .expect("write manifest");
-    let payload_path = temp.path().join("payload.json");
-    std::fs::write(&payload_path, r#"{"event":"created"}"#).expect("write payload");
-
-    let mut cli = test_cli();
-    cli.extension_exec_manifest = Some(manifest_path);
-    cli.extension_exec_hook = Some("run-start".to_string());
-    cli.extension_exec_payload_file = Some(payload_path);
-    execute_extension_exec_command(&cli).expect("extension exec should succeed");
-}
-
-#[test]
-fn regression_execute_extension_exec_command_rejects_undeclared_hook() {
-    let temp = tempdir().expect("tempdir");
-    let bin_dir = temp.path().join("bin");
-    std::fs::create_dir_all(&bin_dir).expect("create bin directory");
-    let script_path = bin_dir.join("hook.sh");
-    std::fs::write(
-        &script_path,
-        "#!/usr/bin/env bash\nread -r _input\nprintf '{\"ok\":true}'\n",
-    )
-    .expect("write hook script");
-    make_script_executable(&script_path);
-    let manifest_path = temp.path().join("extension.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "id": "issue-assistant",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "bin/hook.sh",
-  "hooks": ["run-end"],
-  "permissions": ["run-commands"],
-  "timeout_ms": 5000
-}"#,
-    )
-    .expect("write manifest");
-    let payload_path = temp.path().join("payload.json");
-    std::fs::write(&payload_path, r#"{"event":"created"}"#).expect("write payload");
-
-    let mut cli = test_cli();
-    cli.extension_exec_manifest = Some(manifest_path);
-    cli.extension_exec_hook = Some("run-start".to_string());
-    cli.extension_exec_payload_file = Some(payload_path);
-    let error = execute_extension_exec_command(&cli).expect_err("undeclared hook should fail");
-    assert!(error.to_string().contains("does not declare hook"));
-}
-
-#[test]
-fn regression_execute_extension_exec_command_enforces_timeout() {
-    let temp = tempdir().expect("tempdir");
-    let bin_dir = temp.path().join("bin");
-    std::fs::create_dir_all(&bin_dir).expect("create bin directory");
-    let script_path = bin_dir.join("slow.sh");
-    std::fs::write(
-        &script_path,
-        "#!/usr/bin/env bash\nsleep 1\nprintf '{\"ok\":true}'\n",
-    )
-    .expect("write hook script");
-    make_script_executable(&script_path);
-    let manifest_path = temp.path().join("extension.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "id": "issue-assistant",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "bin/slow.sh",
-  "hooks": ["run-start"],
-  "permissions": ["run-commands"],
-  "timeout_ms": 20
-}"#,
-    )
-    .expect("write manifest");
-    let payload_path = temp.path().join("payload.json");
-    std::fs::write(&payload_path, r#"{"event":"created"}"#).expect("write payload");
-
-    let mut cli = test_cli();
-    cli.extension_exec_manifest = Some(manifest_path);
-    cli.extension_exec_hook = Some("run-start".to_string());
-    cli.extension_exec_payload_file = Some(payload_path);
-    let error = execute_extension_exec_command(&cli).expect_err("timeout should fail");
-    assert!(error.to_string().contains("timed out"));
-}
-
-#[test]
-fn regression_execute_extension_exec_command_rejects_invalid_json_response() {
-    let temp = tempdir().expect("tempdir");
-    let bin_dir = temp.path().join("bin");
-    std::fs::create_dir_all(&bin_dir).expect("create bin directory");
-    let script_path = bin_dir.join("bad.sh");
-    std::fs::write(
-        &script_path,
-        "#!/usr/bin/env bash\nread -r _input\nprintf 'not-json'\n",
-    )
-    .expect("write hook script");
-    make_script_executable(&script_path);
-    let manifest_path = temp.path().join("extension.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "id": "issue-assistant",
-  "version": "0.1.0",
-  "runtime": "process",
-  "entrypoint": "bin/bad.sh",
-  "hooks": ["run-start"],
-  "permissions": ["run-commands"],
-  "timeout_ms": 5000
-}"#,
-    )
-    .expect("write manifest");
-    let payload_path = temp.path().join("payload.json");
-    std::fs::write(&payload_path, r#"{"event":"created"}"#).expect("write payload");
-
-    let mut cli = test_cli();
-    cli.extension_exec_manifest = Some(manifest_path);
-    cli.extension_exec_hook = Some("run-start".to_string());
-    cli.extension_exec_payload_file = Some(payload_path);
-    let error =
-        execute_extension_exec_command(&cli).expect_err("invalid output should be rejected");
-    assert!(error.to_string().contains("response must be valid JSON"));
-}
-
-#[test]
-fn functional_execute_package_validate_command_succeeds_for_valid_manifest() {
-    let temp = tempdir().expect("tempdir");
-    let manifest_path = temp.path().join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut cli = test_cli();
-    cli.package_validate = Some(manifest_path);
-    execute_package_validate_command(&cli).expect("package validate should succeed");
-}
-
-#[test]
-fn regression_execute_package_validate_command_rejects_invalid_manifest() {
-    let temp = tempdir().expect("tempdir");
-    let manifest_path = temp.path().join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 9,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut cli = test_cli();
-    cli.package_validate = Some(manifest_path);
-    let error = execute_package_validate_command(&cli).expect_err("invalid schema should fail");
-    assert!(error
-        .to_string()
-        .contains("unsupported package manifest schema"));
-}
-
-#[test]
-fn functional_execute_package_show_command_succeeds_for_valid_manifest() {
-    let temp = tempdir().expect("tempdir");
-    let manifest_path = temp.path().join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}],
-  "skills": [{"id":"checks","path":"skills/checks/SKILL.md"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut cli = test_cli();
-    cli.package_show = Some(manifest_path);
-    execute_package_show_command(&cli).expect("package show should succeed");
-}
-
-#[test]
-fn regression_execute_package_show_command_rejects_invalid_manifest() {
-    let temp = tempdir().expect("tempdir");
-    let manifest_path = temp.path().join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "invalid",
-  "templates": [{"id":"review","path":"templates/review.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut cli = test_cli();
-    cli.package_show = Some(manifest_path);
-    let error = execute_package_show_command(&cli).expect_err("invalid version should fail");
-    assert!(error.to_string().contains("must follow x.y.z"));
-}
-
-#[test]
-fn functional_execute_package_install_command_succeeds_for_valid_manifest() {
-    let temp = tempdir().expect("tempdir");
-    let package_root = temp.path().join("bundle");
-    std::fs::create_dir_all(package_root.join("templates")).expect("create templates dir");
-    std::fs::write(package_root.join("templates/review.txt"), "template")
-        .expect("write template source");
-
-    let manifest_path = package_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let install_root = temp.path().join("installed");
-    let mut cli = test_cli();
-    cli.package_install = Some(manifest_path);
-    cli.package_install_root = install_root.clone();
-
-    execute_package_install_command(&cli).expect("package install should succeed");
-    assert!(install_root
-        .join("starter-bundle/1.0.0/templates/review.txt")
-        .exists());
-}
-
-#[test]
-fn functional_execute_package_install_command_supports_remote_sources_with_checksum() {
-    let server = MockServer::start();
-    let remote_body = "remote template body";
-    let remote_mock = server.mock(|when, then| {
-        when.method(GET).path("/templates/review.txt");
-        then.status(200).body(remote_body);
-    });
-
-    let temp = tempdir().expect("tempdir");
-    let package_root = temp.path().join("bundle");
-    std::fs::create_dir_all(&package_root).expect("create package root");
-    let checksum = format!("{:x}", Sha256::digest(remote_body.as_bytes()));
-    let manifest_path = package_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        format!(
-            r#"{{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{{
-    "id":"review",
-    "path":"templates/review.txt",
-    "url":"{}/templates/review.txt",
-    "sha256":"sha256:{checksum}"
-  }}]
-}}"#,
-            server.base_url()
-        ),
-    )
-    .expect("write manifest");
-
-    let install_root = temp.path().join("installed");
-    let mut cli = test_cli();
-    cli.package_install = Some(manifest_path);
-    cli.package_install_root = install_root.clone();
-    execute_package_install_command(&cli).expect("package install should succeed");
-    assert_eq!(
-        std::fs::read_to_string(install_root.join("starter-bundle/1.0.0/templates/review.txt"))
-            .expect("read installed template"),
-        remote_body
-    );
-    remote_mock.assert();
-}
-
-#[test]
-fn regression_execute_package_install_command_rejects_missing_component_source() {
-    let temp = tempdir().expect("tempdir");
-    let package_root = temp.path().join("bundle");
-    std::fs::create_dir_all(package_root.join("templates")).expect("create templates dir");
-
-    let manifest_path = package_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/missing.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut cli = test_cli();
-    cli.package_install = Some(manifest_path);
-    cli.package_install_root = temp.path().join("installed");
-    let error = execute_package_install_command(&cli).expect_err("missing source should fail");
-    assert!(error.to_string().contains("does not exist"));
-}
-
-#[test]
-fn regression_execute_package_install_command_rejects_remote_checksum_mismatch() {
-    let server = MockServer::start();
-    let remote_mock = server.mock(|when, then| {
-        when.method(GET).path("/templates/review.txt");
-        then.status(200).body("remote template");
-    });
-
-    let temp = tempdir().expect("tempdir");
-    let package_root = temp.path().join("bundle");
-    std::fs::create_dir_all(&package_root).expect("create package root");
-    let manifest_path = package_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        format!(
-            r#"{{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{{
-    "id":"review",
-    "path":"templates/review.txt",
-    "url":"{}/templates/review.txt",
-    "sha256":"sha256:{}"
-  }}]
-}}"#,
-            server.base_url(),
-            "0".repeat(64)
-        ),
-    )
-    .expect("write manifest");
-
-    let mut cli = test_cli();
-    cli.package_install = Some(manifest_path);
-    cli.package_install_root = temp.path().join("installed");
-    let error =
-        execute_package_install_command(&cli).expect_err("checksum mismatch should fail install");
-    assert!(error.to_string().contains("checksum mismatch"));
-    remote_mock.assert();
-}
-
-#[test]
-fn regression_execute_package_install_command_rejects_unsigned_when_required() {
-    let temp = tempdir().expect("tempdir");
-    let package_root = temp.path().join("bundle");
-    std::fs::create_dir_all(package_root.join("templates")).expect("create templates dir");
-    std::fs::write(package_root.join("templates/review.txt"), "template")
-        .expect("write template source");
-    let manifest_path = package_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut cli = test_cli();
-    cli.package_install = Some(manifest_path);
-    cli.package_install_root = temp.path().join("installed");
-    cli.require_signed_packages = true;
-    let error =
-        execute_package_install_command(&cli).expect_err("unsigned package should fail policy");
-    assert!(error
-        .to_string()
-        .contains("must include signing_key and signature_file"));
-}
-
-#[test]
-fn functional_execute_package_update_command_updates_existing_package() {
-    let temp = tempdir().expect("tempdir");
-    let package_root = temp.path().join("bundle");
-    std::fs::create_dir_all(package_root.join("templates")).expect("create templates dir");
-    let template_path = package_root.join("templates/review.txt");
-    std::fs::write(&template_path, "template-v1").expect("write template source");
-    let manifest_path = package_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let install_root = temp.path().join("installed");
-    let mut install_cli = test_cli();
-    install_cli.package_install = Some(manifest_path.clone());
-    install_cli.package_install_root = install_root.clone();
-    execute_package_install_command(&install_cli).expect("package install should succeed");
-
-    std::fs::write(&template_path, "template-v2").expect("update template source");
-    let mut update_cli = test_cli();
-    update_cli.package_update = Some(manifest_path);
-    update_cli.package_update_root = install_root.clone();
-    execute_package_update_command(&update_cli).expect("package update should succeed");
-    assert_eq!(
-        std::fs::read_to_string(install_root.join("starter-bundle/1.0.0/templates/review.txt"))
-            .expect("read updated template"),
-        "template-v2"
-    );
-}
-
-#[test]
-fn regression_execute_package_update_command_rejects_missing_target() {
-    let temp = tempdir().expect("tempdir");
-    let package_root = temp.path().join("bundle");
-    std::fs::create_dir_all(package_root.join("templates")).expect("create templates dir");
-    std::fs::write(package_root.join("templates/review.txt"), "template")
-        .expect("write template source");
-    let manifest_path = package_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut cli = test_cli();
-    cli.package_update = Some(manifest_path);
-    cli.package_update_root = temp.path().join("installed");
-    let error = execute_package_update_command(&cli).expect_err("missing target should fail");
-    assert!(error.to_string().contains("is not installed"));
-}
-
-#[test]
-fn functional_execute_package_list_command_reports_installed_packages() {
-    let temp = tempdir().expect("tempdir");
-    let package_root = temp.path().join("bundle");
-    std::fs::create_dir_all(package_root.join("templates")).expect("create templates dir");
-    std::fs::write(package_root.join("templates/review.txt"), "template")
-        .expect("write template source");
-    let manifest_path = package_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let install_root = temp.path().join("installed");
-    let mut install_cli = test_cli();
-    install_cli.package_install = Some(manifest_path);
-    install_cli.package_install_root = install_root.clone();
-    execute_package_install_command(&install_cli).expect("package install should succeed");
-
-    let mut list_cli = test_cli();
-    list_cli.package_list = true;
-    list_cli.package_list_root = install_root;
-    execute_package_list_command(&list_cli).expect("package list should succeed");
-}
-
-#[test]
-fn regression_execute_package_list_command_rejects_non_directory_root() {
-    let temp = tempdir().expect("tempdir");
-    let root_file = temp.path().join("not-a-directory.txt");
-    std::fs::write(&root_file, "file root").expect("write root file");
-
-    let mut cli = test_cli();
-    cli.package_list = true;
-    cli.package_list_root = root_file;
-    let error = execute_package_list_command(&cli).expect_err("non-directory root should fail");
-    assert!(error.to_string().contains("is not a directory"));
-}
-
-#[test]
-fn functional_execute_package_remove_command_removes_installed_package() {
-    let temp = tempdir().expect("tempdir");
-    let package_root = temp.path().join("bundle");
-    std::fs::create_dir_all(package_root.join("templates")).expect("create templates dir");
-    std::fs::write(package_root.join("templates/review.txt"), "template")
-        .expect("write template source");
-    let manifest_path = package_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let install_root = temp.path().join("installed");
-    let mut install_cli = test_cli();
-    install_cli.package_install = Some(manifest_path);
-    install_cli.package_install_root = install_root.clone();
-    execute_package_install_command(&install_cli).expect("package install should succeed");
-
-    let mut remove_cli = test_cli();
-    remove_cli.package_remove = Some("starter-bundle@1.0.0".to_string());
-    remove_cli.package_remove_root = install_root.clone();
-    execute_package_remove_command(&remove_cli).expect("package remove should succeed");
-    assert!(!install_root.join("starter-bundle/1.0.0").exists());
-}
-
-#[test]
-fn regression_execute_package_remove_command_rejects_invalid_coordinate() {
-    let mut cli = test_cli();
-    cli.package_remove = Some("starter-bundle".to_string());
-    cli.package_remove_root = PathBuf::from(".tau/packages");
-    let error =
-        execute_package_remove_command(&cli).expect_err("invalid coordinate format should fail");
-    assert!(error.to_string().contains("must follow <name>@<version>"));
-}
-
-#[test]
-fn functional_execute_package_rollback_command_removes_non_target_versions() {
-    let temp = tempdir().expect("tempdir");
-    let install_root = temp.path().join("installed");
-    let install_version = |version: &str, body: &str| {
-        let source_root = temp.path().join(format!("bundle-{version}"));
-        std::fs::create_dir_all(source_root.join("templates")).expect("create templates dir");
-        std::fs::write(source_root.join("templates/review.txt"), body)
-            .expect("write template source");
-        let manifest_path = source_root.join("package.json");
-        std::fs::write(
-            &manifest_path,
-            format!(
-                r#"{{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "{version}",
-  "templates": [{{"id":"review","path":"templates/review.txt"}}]
-}}"#
-            ),
-        )
-        .expect("write manifest");
-
-        let mut install_cli = test_cli();
-        install_cli.package_install = Some(manifest_path);
-        install_cli.package_install_root = install_root.clone();
-        execute_package_install_command(&install_cli).expect("package install should succeed");
-    };
-
-    install_version("1.0.0", "v1");
-    install_version("2.0.0", "v2");
-
-    let mut rollback_cli = test_cli();
-    rollback_cli.package_rollback = Some("starter-bundle@1.0.0".to_string());
-    rollback_cli.package_rollback_root = install_root.clone();
-    execute_package_rollback_command(&rollback_cli).expect("package rollback should succeed");
-    assert!(install_root.join("starter-bundle/1.0.0").exists());
-    assert!(!install_root.join("starter-bundle/2.0.0").exists());
-}
-
-#[test]
-fn regression_execute_package_rollback_command_rejects_missing_target() {
-    let temp = tempdir().expect("tempdir");
-    let mut cli = test_cli();
-    cli.package_rollback = Some("starter-bundle@1.0.0".to_string());
-    cli.package_rollback_root = temp.path().join("installed");
-    let error =
-        execute_package_rollback_command(&cli).expect_err("missing target version should fail");
-    assert!(error.to_string().contains("is not installed"));
-}
-
-#[test]
-fn regression_execute_package_rollback_command_rejects_invalid_coordinate() {
-    let mut cli = test_cli();
-    cli.package_rollback = Some("../starter@1.0.0".to_string());
-    cli.package_rollback_root = PathBuf::from(".tau/packages");
-    let error = execute_package_rollback_command(&cli).expect_err("invalid coordinate should fail");
-    assert!(error
-        .to_string()
-        .contains("must not contain path separators"));
-}
-
-#[test]
-fn functional_execute_package_conflicts_command_reports_detected_collisions() {
-    let temp = tempdir().expect("tempdir");
-    let install_root = temp.path().join("installed");
-    let install_package = |name: &str, body: &str| {
-        let source_root = temp.path().join(format!("bundle-{name}"));
-        std::fs::create_dir_all(source_root.join("templates")).expect("create templates dir");
-        std::fs::write(source_root.join("templates/review.txt"), body)
-            .expect("write template source");
-        let manifest_path = source_root.join("package.json");
-        std::fs::write(
-            &manifest_path,
-            format!(
-                r#"{{
-  "schema_version": 1,
-  "name": "{name}",
-  "version": "1.0.0",
-  "templates": [{{"id":"review","path":"templates/review.txt"}}]
-}}"#
-            ),
-        )
-        .expect("write manifest");
-        let mut install_cli = test_cli();
-        install_cli.package_install = Some(manifest_path);
-        install_cli.package_install_root = install_root.clone();
-        execute_package_install_command(&install_cli).expect("package install should succeed");
-    };
-
-    install_package("alpha", "alpha body");
-    install_package("zeta", "zeta body");
-
-    let mut cli = test_cli();
-    cli.package_conflicts = true;
-    cli.package_conflicts_root = install_root;
-    execute_package_conflicts_command(&cli).expect("package conflicts should succeed");
-}
-
-#[test]
-fn regression_execute_package_conflicts_command_rejects_non_directory_root() {
-    let temp = tempdir().expect("tempdir");
-    let root_file = temp.path().join("not-a-directory.txt");
-    std::fs::write(&root_file, "file root").expect("write root file");
-
-    let mut cli = test_cli();
-    cli.package_conflicts = true;
-    cli.package_conflicts_root = root_file;
-    let error =
-        execute_package_conflicts_command(&cli).expect_err("non-directory root should fail");
-    assert!(error.to_string().contains("is not a directory"));
-}
-
-#[test]
-fn functional_execute_package_activate_command_materializes_components() {
-    let temp = tempdir().expect("tempdir");
-    let install_root = temp.path().join("installed");
-    let source_root = temp.path().join("bundle");
-    std::fs::create_dir_all(source_root.join("templates")).expect("create templates dir");
-    std::fs::create_dir_all(source_root.join("skills/checks")).expect("create skills dir");
-    std::fs::write(source_root.join("templates/review.txt"), "template body")
-        .expect("write template source");
-    std::fs::write(source_root.join("skills/checks/SKILL.md"), "# checks")
-        .expect("write skill source");
-    let manifest_path = source_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "templates": [{"id":"review","path":"templates/review.txt"}],
-  "skills": [{"id":"checks","path":"skills/checks/SKILL.md"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut install_cli = test_cli();
-    install_cli.package_install = Some(manifest_path);
-    install_cli.package_install_root = install_root.clone();
-    execute_package_install_command(&install_cli).expect("package install should succeed");
-
-    let destination_root = temp.path().join("activated");
-    let mut cli = test_cli();
-    cli.package_activate = true;
-    cli.package_activate_root = install_root;
-    cli.package_activate_destination = destination_root.clone();
-    execute_package_activate_command(&cli).expect("package activate should succeed");
-    assert_eq!(
-        std::fs::read_to_string(destination_root.join("templates/review.txt"))
-            .expect("read activated template"),
-        "template body"
-    );
-    assert_eq!(
-        std::fs::read_to_string(destination_root.join("skills/checks/SKILL.md"))
-            .expect("read activated skill"),
-        "# checks"
-    );
-}
-
-#[test]
-fn regression_execute_package_activate_command_rejects_unsupported_conflict_policy() {
-    let mut cli = test_cli();
-    cli.package_activate = true;
-    cli.package_activate_conflict_policy = "unsupported".to_string();
-    let error = execute_package_activate_command(&cli)
-        .expect_err("unsupported conflict policy should fail");
-    assert!(error
-        .to_string()
-        .contains("unsupported package activation conflict policy"));
-}
-
-#[test]
-fn regression_execute_package_activate_on_startup_is_noop_when_disabled() {
-    let temp = tempdir().expect("tempdir");
-    let destination_root = temp.path().join("activated");
-    let mut cli = test_cli();
-    cli.package_activate_root = temp.path().join("installed");
-    cli.package_activate_destination = destination_root.clone();
-    let report = execute_package_activate_on_startup(&cli)
-        .expect("startup activation should allow disabled mode");
-    assert!(report.is_none());
-    assert!(!destination_root.exists());
-}
-
-#[test]
-fn functional_execute_package_activate_on_startup_creates_runtime_skill_alias() {
-    let temp = tempdir().expect("tempdir");
-    let install_root = temp.path().join("installed");
-    let source_root = temp.path().join("bundle");
-    std::fs::create_dir_all(source_root.join("skills/checks")).expect("create skills dir");
-    std::fs::write(source_root.join("skills/checks/SKILL.md"), "# checks")
-        .expect("write skill source");
-    let manifest_path = source_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "skills": [{"id":"checks","path":"skills/checks/SKILL.md"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut install_cli = test_cli();
-    install_cli.package_install = Some(manifest_path);
-    install_cli.package_install_root = install_root.clone();
-    execute_package_install_command(&install_cli).expect("package install should succeed");
-
-    let destination_root = temp.path().join("activated");
-    let mut cli = test_cli();
-    cli.package_activate_on_startup = true;
-    cli.package_activate_root = install_root;
-    cli.package_activate_destination = destination_root.clone();
-    let report = execute_package_activate_on_startup(&cli)
-        .expect("startup activation should succeed")
-        .expect("startup activation should return report");
-    assert_eq!(report.activated_components, 1);
-    assert_eq!(
-        std::fs::read_to_string(destination_root.join("skills/checks/SKILL.md"))
-            .expect("read activated nested skill"),
-        "# checks"
-    );
-    assert_eq!(
-        std::fs::read_to_string(destination_root.join("skills/checks.md"))
-            .expect("read activated skill alias"),
-        "# checks"
-    );
-}
-
-#[test]
-fn integration_compose_startup_system_prompt_uses_activated_skill_aliases() {
-    let temp = tempdir().expect("tempdir");
-    let install_root = temp.path().join("installed");
-    let source_root = temp.path().join("bundle");
-    std::fs::create_dir_all(source_root.join("skills/checks")).expect("create skills dir");
-    std::fs::write(
-        source_root.join("skills/checks/SKILL.md"),
-        "Always run tests",
-    )
-    .expect("write skill source");
-    let manifest_path = source_root.join("package.json");
-    std::fs::write(
-        &manifest_path,
-        r#"{
-  "schema_version": 1,
-  "name": "starter-bundle",
-  "version": "1.0.0",
-  "skills": [{"id":"checks","path":"skills/checks/SKILL.md"}]
-}"#,
-    )
-    .expect("write manifest");
-
-    let mut install_cli = test_cli();
-    install_cli.package_install = Some(manifest_path);
-    install_cli.package_install_root = install_root.clone();
-    execute_package_install_command(&install_cli).expect("package install should succeed");
-
-    let destination_root = temp.path().join("activated");
-    let mut activation_cli = test_cli();
-    activation_cli.package_activate_on_startup = true;
-    activation_cli.package_activate_root = install_root;
-    activation_cli.package_activate_destination = destination_root.clone();
-    execute_package_activate_on_startup(&activation_cli)
-        .expect("startup activation should succeed");
-
-    let mut cli = test_cli();
-    cli.system_prompt = "base prompt".to_string();
-    cli.skills = vec!["checks".to_string()];
-    let composed = compose_startup_system_prompt(&cli, &destination_root.join("skills"))
-        .expect("compose startup prompt");
-    assert!(composed.contains("base prompt"));
-    assert!(composed.contains("Always run tests"));
-}
-
-#[test]
-fn regression_execute_package_activate_on_startup_rejects_unsupported_conflict_policy() {
-    let mut cli = test_cli();
-    cli.package_activate_on_startup = true;
-    cli.package_activate_conflict_policy = "unsupported".to_string();
-    let error = execute_package_activate_on_startup(&cli)
-        .expect_err("unsupported conflict policy should fail");
-    assert!(error
-        .to_string()
-        .contains("unsupported package activation conflict policy"));
-}
+mod extensions_and_packages;
